chore: split test files for transforms
