+++ packages/compiler-vapor/__tests__/compile.test.ts
@@ -1,9 +1,4 @@
-import {
-  type RootNode,
-  BindingTypes,
-  ErrorCodes,
-  DOMErrorCodes,
-} from '@vue/compiler-dom'
+import { type RootNode, BindingTypes } from '@vue/compiler-dom'
 import { type CompilerOptions, compile as _compile } from '../src'
 
 function compile(template: string | RootNode, options: CompilerOptions = {}) {
@@ -61,222 +56,6 @@ describe('compile', () => {
   })
 
   describe('directives', () => {
-    describe('v-bind', () => {
-      test('simple expression', async () => {
-        const code = await compile(`<div :id="id"></div>`, {
-          bindingMetadata: {
-            id: BindingTypes.SETUP_REF,
-          },
-        })
-        expect(code).matchSnapshot()
-      })
-
-      test('should error if no expression', async () => {
-        const onError = vi.fn()
-        const code = await compile(`<div v-bind:arg="" />`, { onError })
-
-        expect(onError.mock.calls[0][0]).toMatchObject({
-          code: ErrorCodes.X_V_BIND_NO_EXPRESSION,
-          loc: {
-            start: {
-              line: 1,
-              column: 6,
-            },
-            end: {
-              line: 1,
-              column: 19,
-            },
-          },
-        })
-
-        expect(code).matchSnapshot()
-        // the arg is static
-        expect(code).contains(JSON.stringify('<div arg=""></div>'))
-      })
-
-      test('no expression', async () => {
-        const code = await compile('<div v-bind:id />', {
-          bindingMetadata: {
-            id: BindingTypes.SETUP_REF,
-          },
-        })
-
-        expect(code).matchSnapshot()
-        expect(code).contains('_setAttr(n1, "id", undefined, _ctx.id)')
-      })
-
-      test('no expression (shorthand)', async () => {
-        const code = await compile('<div :camel-case />', {
-          bindingMetadata: {
-            camelCase: BindingTypes.SETUP_REF,
-          },
-        })
-
-        expect(code).matchSnapshot()
-        expect(code).contains(
-          '_setAttr(n1, "camel-case", undefined, _ctx.camelCase)',
-        )
-      })
-
-      test('dynamic arg', async () => {
-        const code = await compile('<div v-bind:[id]="id"/>', {
-          bindingMetadata: {
-            id: BindingTypes.SETUP_REF,
-          },
-        })
-
-        expect(code).matchSnapshot()
-        expect(code).contains('_setAttr(n1, _ctx.id, undefined, _ctx.id)')
-      })
-
-      // TODO: camel modifier for v-bind
-      test.fails('.camel modifier', async () => {
-        const code = await compile(`<div v-bind:foo-bar.camel="id"/>`)
-
-        expect(code).matchSnapshot()
-        expect(code).contains('fooBar')
-      })
-    })
-
-    describe('v-on', () => {
-      test('simple expression', async () => {
-        const code = await compile(`<div @click="handleClick"></div>`, {
-          bindingMetadata: {
-            handleClick: BindingTypes.SETUP_CONST,
-          },
-        })
-        expect(code).matchSnapshot()
-      })
-
-      test('should error if no expression AND no modifier', async () => {
-        const onError = vi.fn()
-        await compile(`<div v-on:click />`, { onError })
-        expect(onError.mock.calls[0][0]).toMatchObject({
-          code: ErrorCodes.X_V_ON_NO_EXPRESSION,
-          loc: {
-            start: {
-              line: 1,
-              column: 6,
-            },
-            end: {
-              line: 1,
-              column: 16,
-            },
-          },
-        })
-      })
-
-      test('event modifier', async () => {
-        const code = await compile(
-          `<a @click.stop="handleEvent"></a>
-            <form @submit.prevent="handleEvent"></form>
-            <a @click.stop.prevent="handleEvent"></a>
-            <div @click.self="handleEvent"></div>
-            <div @click.capture="handleEvent"></div>
-            <a @click.once="handleEvent"></a>
-            <div @scroll.passive="handleEvent"></div>
-            <input @click.right="handleEvent" />
-            <input @click.left="handleEvent" />
-            <input @click.middle="handleEvent" />
-            <input @click.enter.right="handleEvent" />
-            <input @keyup.enter="handleEvent" />
-            <input @keyup.tab="handleEvent" />
-            <input @keyup.delete="handleEvent" />
-            <input @keyup.esc="handleEvent" />
-            <input @keyup.space="handleEvent" />
-            <input @keyup.up="handleEvent" />
-            <input @keyup.down="handleEvent" />
-            <input @keyup.left="handleEvent" />
-            <input @keyup.middle="submit" />
-            <input @keyup.middle.self="submit" />
-            <input @keyup.self.enter="handleEvent" />`,
-          {
-            bindingMetadata: {
-              handleEvent: BindingTypes.SETUP_CONST,
-            },
-          },
-        )
-        expect(code).matchSnapshot()
-      })
-    })
-
-    describe('v-html', () => {
-      test('simple expression', async () => {
-        const code = await compile(`<div v-html="code"></div>`, {
-          bindingMetadata: {
-            code: BindingTypes.SETUP_REF,
-          },
-        })
-        expect(code).matchSnapshot()
-      })
-
-      test('should raise error and ignore children when v-html is present', async () => {
-        const onError = vi.fn()
-        const code = await compile(`<div v-html="test">hello</div>`, {
-          onError,
-        })
-        expect(code).matchSnapshot()
-        expect(onError.mock.calls).toMatchObject([
-          [{ code: DOMErrorCodes.X_V_HTML_WITH_CHILDREN }],
-        ])
-      })
-
-      test('should raise error if has no expression', async () => {
-        const onError = vi.fn()
-        const code = await compile(`<div v-html></div>`, {
-          onError,
-        })
-        expect(code).matchSnapshot()
-        expect(onError.mock.calls).toMatchObject([
-          [{ code: DOMErrorCodes.X_V_HTML_NO_EXPRESSION }],
-        ])
-      })
-    })
-
-    describe('v-text', () => {
-      test('simple expression', async () => {
-        const code = await compile(`<div v-text="str"></div>`, {
-          bindingMetadata: {
-            str: BindingTypes.SETUP_REF,
-          },
-        })
-        expect(code).matchSnapshot()
-      })
-
-      test('no expression', async () => {
-        const onError = vi.fn()
-        const code = await compile(`<div v-text></div>`, { onError })
-        expect(code).matchSnapshot()
-        expect(onError.mock.calls).toMatchObject([
-          [{ code: DOMErrorCodes.X_V_TEXT_NO_EXPRESSION }],
-        ])
-      })
-    })
-
-    describe('v-once', () => {
-      test('basic', async () => {
-        const code = await compile(
-          `<div v-once>
-            {{ msg }}
-            <span :class="clz" />
-          </div>`,
-          {
-            bindingMetadata: {
-              msg: BindingTypes.SETUP_REF,
-              clz: BindingTypes.SETUP_REF,
-            },
-          },
-        )
-        expect(code).matchSnapshot()
-      })
-
-      test('as root node', async () => {
-        const code = await compile(`<div :id="foo" v-once />`)
-        expect(code).toMatchSnapshot()
-        expect(code).not.contains('effect')
-      })
-    })
-
     describe('v-pre', () => {
       test('basic', async () => {
         const code = await compile(
