+++ packages/compiler-vapor/__tests__/transforms/transformInterpolation.test.ts
@@ -1,0 +1,2 @@
+// TODO: add tests for this transform
+test('baisc', () => {})
